refactor: fix race conditions, add accessors, split handlers, improve reliability
Phase 1 — Critical data race fixes:
- Fix GetPlayer() TOCTOU: replace sync.Mutex with sync.RWMutex for concurrent-safe map reads
- Fix handleVoiceRecoveryFailure: protect VoiceConnection/VoiceChannelID writes with VoiceChannelMutex
- Add controller/accessors.go: GetCurrentSong(), GetCurrentItem(), GetVoiceChannelID(),
  GetVoiceConnection(), Get/SetLastTextChannelID(), GetQueueSnapshot(), ShouldJoinVoice()
- Fix all 13+ raw CurrentSong accesses in handlers — eliminates nil-dereference crash paths
- Fix shouldJoin pattern in handlers: use ShouldJoinVoice() accessor (3 locations)
- Fix Queue.Items unlocked iteration in onView: use GetQueueSnapshot()
- Replace WaitForStreamURL 300x100ms busy-poll with a channel on GuildQueueItem

Phase 2 — Correctness and reliability:
- Add playerCtx/playerCancel to GuildPlayer: Reset() cancels in-flight voice recovery retries
- Fix Reset() overly-wide Queue.Mutex critical section — only state mutation is locked now
- Fix Loader.Cancel() dropped-cancel window: buffer canceled chan to size 1, drain stale signals
- Fix RemoveFavoriteByIndex TOCTOU: wrap SELECT+DELETE in a transaction
- Fix SQLite: SetMaxOpenConns(1) + SetMaxIdleConns(1) + PRAGMA busy_timeout=5000
- Fix GetFavorites silent date parse discard: use multi-format parser, log on failure

Phase 3 — Performance:
- Gemini client initialized once at startup (gemini.Init()) instead of per-call
- Dispatch handleAdd to worker goroutine — queue listener no longer blocks on yt-dlp during playlist loads
- Manager constructed once in main.go instead of per HTTP request

Phase 4 — Code organisation:
- Split handlers/handlers.go (2439 lines) into 8 focused files: queue.go, playback.go,
  spotify.go, applemusic.go, youtube.go, database.go, misc.go
- Consolidate duplicate ExtractArtist / ExtractArtistFromTitle into discord/embeds.go
- Remove dead code: printResponse() in gemini.go, player.completed channel (never sent to)
  and its unreachable select branch; simplify safeSendOpus accordingly

Phase 5 — Minor cleanup:
- Remove deprecated rand.Seed call in hints.go (auto-seeded since Go 1.20)
- Add UNIQUE index on user_favorites(user_id, guild_id, video_id) as a new migration
